pyproject.toml
[tool.poetry]
name = "mazes4prog"
version = "0.1.0"
description = ""
authors = ["aporter <aporter@blueorigin.com>"]

[tool.poetry.dependencies]
python = "^3.8"
Pillow = "^8.4.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
